Match HSD dobj

+++ src/sysdolphin/pobj.c
@@ -1,6 +1,6 @@
 #include "pobj.h"
 
-u16 HSD_PObjGetFlags(HSD_PObj* pobj)
+u32 HSD_PObjGetFlags(HSD_PObj* pobj)
 {
     if (pobj != NULL) {
         return pobj->flags;
